Kernel: Xiaomi kernel changes for Redmi Note 10 Android R
The Patch based on QualComm release TAG:LA.UM.9.1.r1-07700-SMxxx0.0-1
The kernel config file used is atoll-perf_defconfig

+++ drivers/char/diag/diag_pcie.c
@@ -1,4 +1,5 @@
 /* Copyright (c) 2018-2020, The Linux Foundation. All rights reserved.
+ * Copyright (C) 2021 XiaoMi, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -416,31 +417,41 @@ void diag_pcie_client_cb(struct mhi_dev_client_cb_data *cb_data)
 {
 	struct diag_pcie_info *pcie_info = NULL;
 
-	if (!cb_data)
+	if (!cb_data) {
+		pr_err("diag: %s: Invalid cb_data\n", __func__);
 		return;
-
+	}
 	pcie_info = cb_data->user_data;
-	if (!pcie_info)
+	if (!pcie_info) {
+		pr_err("diag: %s: Invalid pcie_info\n", __func__);
 		return;
-
+	}
 	switch (cb_data->ctrl_info) {
 	case  MHI_STATE_CONNECTED:
 		if (cb_data->channel == pcie_info->out_chan) {
 			DIAG_LOG(DIAG_DEBUG_MUX,
-				" Received connect event from MHI for %d",
+				"diag: Received connect event from MHI for %d\n",
 				pcie_info->out_chan);
-			if (atomic_read(&pcie_info->enabled))
+			if (atomic_read(&pcie_info->enabled)) {
+				DIAG_LOG(DIAG_DEBUG_MUX,
+				"diag: Channel %d is already enabled\n",
+				pcie_info->out_chan);
 				return;
+			}
 			queue_work(pcie_info->wq, &pcie_info->open_work);
 		}
 		break;
 	case MHI_STATE_DISCONNECTED:
 		if (cb_data->channel == pcie_info->out_chan) {
 			DIAG_LOG(DIAG_DEBUG_MUX,
-				" Received disconnect event from MHI for %d",
+				"diag: Received disconnect event from MHI for %d\n",
 				pcie_info->out_chan);
-			if (!atomic_read(&pcie_info->enabled))
+			if (!atomic_read(&pcie_info->enabled)) {
+				DIAG_LOG(DIAG_DEBUG_MUX,
+				"diag: Channel %d is already disabled\n",
+				pcie_info->out_chan);
 				return;
+			}
 			queue_work(pcie_info->wq, &pcie_info->close_work);
 		}
 		break;
@@ -490,12 +501,9 @@ static void diag_pcie_connect(struct diag_pcie_info *ch)
 	queue_work(ch->wq, &(ch->read_work));
 }
 
-void diag_pcie_open_work_fn(struct work_struct *work)
+static void diag_pcie_open_channels(struct diag_pcie_info *pcie_info)
 {
 	int rc = 0;
-	struct diag_pcie_info *pcie_info = container_of(work,
-						      struct diag_pcie_info,
-						      open_work);
 
 	if (!pcie_info || atomic_read(&pcie_info->enabled))
 		return;
@@ -538,6 +546,15 @@ handle_in_err:
 handle_not_rdy_err:
 	mutex_unlock(&pcie_info->in_chan_lock);
 	mutex_unlock(&pcie_info->out_chan_lock);
+}
+
+void diag_pcie_open_work_fn(struct work_struct *work)
+{
+	struct diag_pcie_info *pcie_info = container_of(work,
+						      struct diag_pcie_info,
+						      open_work);
+
+	diag_pcie_open_channels(pcie_info);
 }
 
 /*
@@ -679,6 +696,8 @@ int diag_pcie_register(int id, int ctxt, struct diag_mux_ops *ops)
 		return -EIO;
 	}
 
+	pr_info("diag: Pcie registration initiated for id: %d\n", id);
+
 	ch = &diag_pcie[id];
 	ch->ops = ops;
 	ch->ctxt = ctxt;
@@ -692,17 +711,27 @@ int diag_pcie_register(int id, int ctxt, struct diag_mux_ops *ops)
 	strlcpy(wq_name, "DIAG_PCIE_", sizeof(wq_name));
 	strlcat(wq_name, ch->name, sizeof(wq_name));
 	ch->wq = create_singlethread_workqueue(wq_name);
-	if (!ch->wq)
+	if (!ch->wq) {
+		pr_err("diag: %s: failed creating workqueue for wq_name: %s\n",
+		__func__, wq_name);
 		return -ENOMEM;
+	}
+	DIAG_LOG(DIAG_DEBUG_MUX, "diag: created wq: %s\n", wq_name);
 	diagmem_init(driver, ch->mempool);
 	mutex_init(&ch->in_chan_lock);
 	mutex_init(&ch->out_chan_lock);
 	rc = diag_register_pcie_channels(ch);
-	if (rc < 0) {
+	if (rc == -EEXIST) {
+		pr_err("diag: Handled -EEXIST error\n");
+		diag_pcie_open_channels(ch);
+	} else if (rc < 0 && rc != -EEXIST) {
 		if (ch->wq)
 			destroy_workqueue(ch->wq);
 		kfree(ch->in_chan_attr.read_buffer);
+		pr_err("diag: %s: failed registering pcie channels\n",
+		__func__);
 		return rc;
 	}
+	pr_info("diag: pcie channel with id: %d registered successfully\n", id);
 	return 0;
 }
